chore(export): (WIP) move tests
