+++ packages/akashic-cli-export/src/zip/getFromHttps.ts
@@ -1,0 +1,17 @@
+import * as https from "https";
+
+export function getFromHttps(url: string): Promise<string> {
+	return new Promise((resolve, reject) => {
+		const request = https.get(url, (res) => {
+			if (res.statusCode! >= 400) { // リファレンスを見ても statusCode がないケースがなさそうなので !
+				return reject(new Error(`Failed to get resource. url: ${url}. status code: ${res.statusCode}.`));
+			}
+			const bodies: string[] = [];
+			res.on("data", (chunk) => {
+				bodies.push(chunk.toString());
+			});
+			res.on("end", () => resolve(bodies.join("")));
+		});
+		request.on("error", (err) => reject(err));
+	});
+}
