+++ packages/akashic-cli-export/src/zip/transformCompleteEnvironment.ts
@@ -1,5 +1,5 @@
-import * as https from "https";
 import { GameConfiguration } from "@akashic/akashic-cli-commons/lib/GameConfiguration";
+import { getFromHttps } from "./getFromHttps";
 
 // TODO game-configuration への移行後、削除する
 interface GameConfigurationEnvrironment {
@@ -47,20 +47,4 @@ export async function transformCompleteEnvrironment(gamejson: GameConfiguration)
 			environment["akashic-runtime"].flavor = "-canvas";
 		}
 	}
-}
-
-export function getFromHttps(url: string): Promise<string> {
-	return new Promise((resolve, reject) => {
-		const request = https.get(url, (res) => {
-			if (res.statusCode! >= 400) { // リファレンスを見ても statusCode がないケースがなさそうなので !
-				return reject(new Error(`Failed to get resource. url: ${url}. status code: ${res.statusCode}.`));
-			}
-			const bodies: string[] = [];
-			res.on("data", (chunk) => {
-				bodies.push(chunk.toString());
-			});
-			res.on("end", () => resolve(bodies.join("")));
-		});
-		request.on("error", (err) => reject(err));
-	});
 }
